// WebApi/ClientApp/src/models/responses/category/IDonViCts.ts
export interface IDonViCts {
    id: number;
    donvi_ma_dv: string;
    url: string;
    file_name: string;
    serial_number: string;
    subject: string;
    issuer: string;
    not_before: string;
    not_after: string;
    signature_algorithm: string;
    version: string;
    serial_type_id: number;
    rs_ma_but_ky: string;
    user_id: number;
    is_active: boolean;
}